♻ REFACTOR: Moved Item structs to schema and flatten data::item module to item
