+++ Cargo.toml
@@ -1,3 +1,8 @@
 [workspace]
 resolver = "2"
 members = ["database", "schema"]
+
+[workspace.dependencies]
+sqlx = { version = "0.7.2" }
+serde = { version = "1.0.173", features = ["derive"] }
+itertools = "0.11.0"
