+++ database/Cargo.toml
@@ -1,18 +1,17 @@
 [package]
 name = "database"
-version = "0.5.2"
+version = "0.6.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+schema = { path = "../schema" }
+serde = { workspace = true }
+itertools = { workspace = true }
+sqlx = { workspace = true }
 async-trait = "0.1.74"
-chrono = { version = "0.4.26", features = ["serde"] }
-csv = "1.2.2"
-itertools = "0.11.0"
-serde = { version = "1.0.173", features = ["derive"] }
-serde_json = "1.0.108"
-sqlx = { version = "0.7.2", features = ["runtime-tokio-rustls", "chrono", "macros", "migrate"] }
+# chrono = { version = "0.4.26", features = ["serde"] }
+# csv = "1.2.2"
 tracing = "0.1.40"
 
 [features]
@@ -22,3 +21,4 @@ postgres = ["sqlx/postgres"]
 [dev-dependencies]
 tokio-test = "0.4.3"
 tracing-test = "0.2.4"
+serde_json = "1.0.108"
