+++ database/src/changes.rs
@@ -2,18 +2,21 @@
 
 use std::fmt::Display;
 
+use crate::engine::{DbEngine, SqlxPool};
+use crate::io::interface::{AttachmentItemMod, Item, ItemMod};
+use crate::items::konsep::KonsepMod;
+
+use schema::items::cakupan::Cakupan as CakupanItem;
+use schema::items::kata_asing::KataAsing as KataAsingItem;
+use schema::items::konsep::Konsep as KonsepItem;
+use schema::items::lemma::Lemma;
+use schema::metatype::AutoGen;
+
 use itertools::Itertools;
 use tracing::instrument;
 
-use crate::engine::{DbEngine, SqlxPool};
-use crate::io::interface::{AttachmentItemMod, Item, ItemMod};
-use crate::{
-    data::{CakupanItem, KonsepItemMod, LemmaItem},
-    prelude::{AutoGen, KataAsingItem, KonsepItem},
-};
-
 /// [ItemMod] fields.
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Clone, PartialEq)]
 pub enum FieldMod<T> {
     New(T),
     Fixed(T),
@@ -42,7 +45,7 @@ impl<T: PartialEq> FieldMod<T> {
 ///
 /// An attachment field is indicated by a vector of children [Items](Item).
 /// The children items in this struct implements [ItemMod].
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Clone, PartialEq)]
 pub struct AttachmentMod<A: ItemMod> {
     pub attached: Vec<A>,
     pub detached: Vec<A>,
@@ -128,31 +131,18 @@ pub trait CompareAttachable<I: PartialEq + Clone + Item<IntoMod = A>, A: ItemMod
     }
 }
 
-impl<I> CompareAttachable<CakupanItem, CakupanItem> for KonsepItem<I> {
-    fn items(&self) -> Vec<CakupanItem> {
-        Vec::clone(&self.cakupans)
-    }
-}
-impl<I> CompareAttachable<KataAsingItem, KataAsingItem> for KonsepItem<I> {
-    fn items(&self) -> Vec<KataAsingItem> {
-        Vec::clone(&self.kata_asing)
-    }
-}
-
-impl<I: Copy + Clone + PartialEq + PartialOrd + Display>
-    CompareAttachable<KonsepItem<I>, KonsepItemMod<I>> for LemmaItem<I>
-{
+impl<I: Copy + Clone + PartialEq> CompareAttachable<KonsepItem<I>, KonsepMod<I>> for Lemma<I> {
     fn items(&self) -> Vec<KonsepItem<I>> {
         Vec::clone(&self.konseps)
     }
-    fn find_attached(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepItemMod<I>> {
+    fn find_attached(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepMod<I>> {
         Vec::clone(other)
             .into_iter()
             .filter(|item| item.id == AutoGen::Unknown)
-            .map(|item| KonsepItemMod::from_item(&item))
+            .map(|item| KonsepMod::from_item(&item))
             .collect_vec()
     }
-    fn find_detached(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepItemMod<I>> {
+    fn find_detached(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepMod<I>> {
         let other_ids = Vec::clone(other)
             .into_iter()
             .filter(|item| item.id != AutoGen::Unknown)
@@ -161,10 +151,10 @@ impl<I: Copy + Clone + PartialEq + PartialOrd + Display>
         Vec::clone(&self.items())
             .into_iter()
             .filter(|item| !other_ids.contains(&item.id))
-            .map(|item| KonsepItemMod::from_item(&item))
+            .map(|item| KonsepMod::from_item(&item))
             .collect_vec()
     }
-    fn find_modified(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepItemMod<I>> {
+    fn find_modified(&self, other: &Vec<KonsepItem<I>>) -> Vec<KonsepMod<I>> {
         let detached_id = self
             .find_detached(other)
             .iter()
@@ -173,11 +163,19 @@ impl<I: Copy + Clone + PartialEq + PartialOrd + Display>
         let old = Vec::clone(&self.items())
             .into_iter()
             .filter(|item| !detached_id.contains(&item.id))
-            .sorted_by(|a, b| a.id.partial_cmp(&b.id).unwrap_or(std::cmp::Ordering::Equal));
+            .sorted_by(|a, b| {
+                a.keterangan
+                    .partial_cmp(&b.keterangan)
+                    .unwrap_or(std::cmp::Ordering::Equal)
+            });
         let new = Vec::clone(other)
             .into_iter()
             .filter(|item| item.id != AutoGen::Unknown)
-            .sorted_by(|a, b| a.id.partial_cmp(&b.id).unwrap_or(std::cmp::Ordering::Equal));
+            .sorted_by(|a, b| {
+                a.keterangan
+                    .partial_cmp(&b.keterangan)
+                    .unwrap_or(std::cmp::Ordering::Equal)
+            });
         dbg!(old.clone().map(|i| i.keterangan).collect_vec());
         dbg!(new.clone().map(|i| i.keterangan).collect_vec());
         assert_eq!(
@@ -193,23 +191,24 @@ impl<I: Copy + Clone + PartialEq + PartialOrd + Display>
 
 #[cfg(test)]
 mod test {
+    // FIX: Assert eq no longer works because we no longer impl Debug
     use super::*;
-    use crate::data::LemmaItemMod;
     use crate::io::interface::{FromView, SubmitMod};
+    use crate::items::lemma::LemmaMod;
     use crate::views::LemmaWithKonsepView;
     use tracing_test::traced_test;
 
     #[sqlx::test]
     #[traced_test]
     fn test_new_lemma(pool: sqlx::Pool<sqlx::Sqlite>) -> Result<(), sqlx::Error> {
-        let new = LemmaItem {
+        let new = Lemma {
             id: AutoGen::Unknown,
             lemma: "cakera tokokan".into(),
             konseps: vec![],
         };
-        LemmaItemMod::from_item(&new).submit_mod(&pool).await?;
+        LemmaMod::from_item(&new).submit_mod(&pool).await?;
         let views = LemmaWithKonsepView::query_lemma("cakera tokokan".into(), &pool).await?;
-        let data = LemmaItem::from_views(&views);
+        let data = Lemma::from_views(&views);
         assert_eq!(data, vec![new]);
         Ok(())
     }
@@ -218,13 +217,13 @@ mod test {
     #[traced_test]
     fn test_diff_handling(pool: sqlx::Pool<sqlx::Sqlite>) -> Result<(), sqlx::Error> {
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         let old = data
             .first()
             .expect("Vec<LemmaDataRepr> is zero sized")
             .to_owned();
         assert_eq!(&old.konseps.len(), &1);
-        let new: LemmaItem = LemmaItem {
+        let new: Lemma = Lemma {
             id: AutoGen::Known(1),
             lemma: "cakera tokokan".into(),
             konseps: vec![
@@ -249,7 +248,7 @@ mod test {
         };
         old.modify_into(&new).unwrap().submit_mod(&pool).await?;
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         assert_eq!(data, vec![new]);
         Ok(())
     }
@@ -257,7 +256,7 @@ mod test {
     #[test]
     #[traced_test]
     fn test_diff_handling_detach_konsep() -> Result<(), Box<dyn std::error::Error>> {
-        let lemma_1 = LemmaItem {
+        let lemma_1 = Lemma {
             id: AutoGen::Known(1),
             lemma: "cubaan".into(),
             konseps: vec![
@@ -284,7 +283,7 @@ mod test {
                 },
             ],
         };
-        let lemma_2 = LemmaItem {
+        let lemma_2 = Lemma {
             id: AutoGen::Known(1),
             lemma: "cubaan".into(),
             konseps: vec![
@@ -307,7 +306,7 @@ mod test {
         let attachment_mod = lemma_1.compare_attachment(lemma_2.konseps);
         assert_eq!(
             attachment_mod.detached,
-            vec![KonsepItemMod {
+            vec![KonsepMod {
                 id: AutoGen::Known(2),
                 keterangan: FieldMod::Fixed("cubaan padam".into()),
                 golongan_kata: FieldMod::Fixed("kata nama".into()),
@@ -318,14 +317,14 @@ mod test {
         assert_eq!(
             attachment_mod.modified,
             vec![
-                KonsepItemMod {
+                KonsepMod {
                     id: AutoGen::Known(1),
                     keterangan: FieldMod::Fixed("cubaan simpan 1/2".into()),
                     golongan_kata: FieldMod::Fixed("kata nama".into()),
                     cakupans: AttachmentMod::empty(),
                     kata_asing: AttachmentMod::empty(),
                 },
-                KonsepItemMod {
+                KonsepMod {
                     id: AutoGen::Known(3),
                     keterangan: FieldMod::Fixed("cubaan simpan 2/2".into()),
                     golongan_kata: FieldMod::Fixed("kata nama".into()),
@@ -343,13 +342,13 @@ mod test {
         pool: sqlx::Pool<sqlx::Sqlite>,
     ) -> Result<(), sqlx::Error> {
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         let old = data
             .first()
             .expect("Vec<LemmaDataRepr> is zero sized")
             .to_owned();
         assert_eq!(&old.konseps.len(), &1);
-        let new: LemmaItem = LemmaItem {
+        let new: Lemma = Lemma {
             id: AutoGen::Known(1),
             lemma: "cakera tokokan".into(),
             konseps: vec![KonsepItem {
@@ -365,7 +364,7 @@ mod test {
         };
         old.modify_into(&new).unwrap().submit_mod(&pool).await?;
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         assert_eq!(
             data.first()
                 .expect("Here?")
@@ -382,13 +381,13 @@ mod test {
         pool: sqlx::Pool<sqlx::Sqlite>,
     ) -> Result<(), sqlx::Error> {
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         let old = data
             .first()
             .expect("Vec<LemmaDataRepr> is zero sized")
             .to_owned();
         assert_eq!(&old.konseps.len(), &1);
-        let new: LemmaItem = LemmaItem {
+        let new: Lemma = Lemma {
             id: AutoGen::Known(1),
             lemma: "cakera tokokan".into(),
             konseps: vec![KonsepItem {
@@ -401,7 +400,7 @@ mod test {
         };
         old.modify_into(&new).unwrap().submit_mod(&pool).await?;
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         assert_eq!(data, vec![new]);
         Ok(())
     }
@@ -410,13 +409,13 @@ mod test {
         pool: sqlx::Pool<sqlx::Sqlite>,
     ) -> Result<(), sqlx::Error> {
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
         let old = data
             .first()
             .expect("Vec<LemmaDataRepr> is zero sized")
             .to_owned();
         assert_eq!(&old.konseps.len(), &2);
-        let new: LemmaItem = LemmaItem {
+        let new: Lemma = Lemma {
             id: AutoGen::Known(1),
             lemma: "cakera tokokan".into(),
             konseps: vec![
@@ -438,7 +437,7 @@ mod test {
         };
         old.modify_into(&new).unwrap().submit_mod(&pool).await?;
         let view = LemmaWithKonsepView::query_all(&pool).await?;
-        let data = LemmaItem::from_views(&view);
+        let data = Lemma::from_views(&view);
 
         assert_eq!(data, vec![new]);
         Ok(())
