+++ database/src/views.rs
@@ -2,20 +2,22 @@
 
 use std::collections::HashMap;
 
-use itertools::Itertools;
-use crate::data::items::konsep::KonsepHashMap;
 use crate::engine::{DbEngine, LemmaWithKonsepViewOption, Query, SqlxPool};
-use crate::io::interface::IntoViewMap;
+use crate::io::interface::{IntoViewMap, View};
+use crate::items::konsep::KonsepHashMap;
 
-/// A [View](crate::io::interface::View) whose query joins konsep to its tags (cakupan, kata_asing)
+use itertools::Itertools;
+
+/// A [View] whose query joins konsep to its tags (cakupan, kata_asing)
 ///
 /// Running this command:
 /// ```no_run
-/// # use database::views::LemmaWithKonsepView;
+/// use database::views::LemmaWithKonsepView;
+/// use database::engine::DbEngine;
 /// #
 /// # tokio_test::block_on(async {
 /// # let pool_url = ":memory:";
-/// let pool = sqlx::SqlitePool::connect(pool_url).await.expect("Connection error");
+/// let pool = DbEngine::sqlite(pool_url).await;
 /// let lemma_views: Vec<LemmaWithKonsepView> = LemmaWithKonsepView::query_lemma(
 ///         "aplikasi".to_string(),
 ///          &pool
@@ -28,25 +30,25 @@ use crate::io::interface::IntoViewMap;
 ///```
 #[derive(Debug, Clone, sqlx::FromRow, Default)]
 pub struct LemmaWithKonsepView {
-    /// ID (Primary): [LemmaItem.id](crate::data::LemmaItem#structfield.id)
+    /// ID (Primary): [Lemma.id](schema::items::lemma::Lemma#structfield.id)
     pub l_id: i64,
-    /// Data (Primary): [LemmaItem.nama](crate::data::LemmaItem#structfield.nama)
+    /// Data (Primary): [Lemma.nama](schema::items::lemma::Lemma#structfield.nama)
     pub lemma: String,
-    /// ID (Secondary): [KonsepItem.id](crate::data::KonsepItem#structfield.id)
+    /// ID (Secondary): [Konsep.id](schema::items::konsep::Konsep#structfield.id)
     pub k_id: i64,
-    /// Data (Secondary): [KonsepItem.keterangan](crate::data::KonsepItem#structfield.keterangan)
+    /// Data (Secondary): [Konsep.keterangan](schema::items::konsep::Konsep#structfield.keterangan)
     pub konsep: Option<String>,
-    /// Data (Secondary): [KonsepItem.golongan_kata](crate::data::KataAsingItem#structfield.golongan_kata)
+    /// Data (Secondary): [Konsep.golongan_kata](schema::items::kata_asing::KataAsing#structfield.golongan_kata)
     pub golongan_kata: Option<String>,
-    /// Attachment: [CakupanItem](crate::data::CakupanItem)
+    /// Attachment: [CakupanItem](schema::items::cakupan::Cakupan)
     pub cakupan: Option<String>,
-    /// Attachment: [KataAsingItem.nama](crate::data::KataAsingItem#structfield.nama)
+    /// Attachment: [KataAsing.nama](schema::items::kata_asing::KataAsing#structfield.nama)
     pub kata_asing: Option<String>,
-    /// Attachment: [KataAsingItem.bahasa](crate::data::KataAsingItem#structfield.bahasa)
+    /// Attachment: [KataAsing.bahasa](schema::items::kata_asing::KataAsing#structfield.bahasa)
     pub bahasa_asing: Option<String>,
 }
 
-impl crate::io::interface::View for LemmaWithKonsepView {
+impl View for LemmaWithKonsepView {
     type SOURCE = ();
 }
 
@@ -55,18 +57,20 @@ impl LemmaWithKonsepView {
     /// Query a single lemma with its associated konseps and attachments.
     pub async fn query_lemma(
         lemma: String,
-        engine: &DbEngine<sqlx::SqlitePool>
-    ) -> sqlx::Result<Vec<Self>>  {
-        engine.select(LemmaWithKonsepViewOption { lemma: Some(lemma) }).await
+        engine: &DbEngine<sqlx::SqlitePool>,
+    ) -> sqlx::Result<Vec<Self>> {
+        engine
+            .select(LemmaWithKonsepViewOption { lemma: Some(lemma) })
+            .await
     }
 
     /// Query all lemmas.
-    pub async fn query_all(
-        engine: &DbEngine<sqlx::SqlitePool>
-    ) -> sqlx::Result<Vec<Self>> {
+    pub async fn query_all(engine: &DbEngine<sqlx::SqlitePool>) -> sqlx::Result<Vec<Self>> {
         // TODO: Might be a good idea to add limit
         // TODO: And maybe sort by reverse chronology
-        engine.select(LemmaWithKonsepViewOption {lemma: None}).await
+        engine
+            .select(LemmaWithKonsepViewOption { lemma: None })
+            .await
     }
 }
 #[cfg(feature = "postgres")]
@@ -74,18 +78,20 @@ impl LemmaWithKonsepView {
     /// Query a single lemma with its associated konseps and attachments.
     pub async fn query_lemma(
         lemma: String,
-        engine: &DbEngine<sqlx::PgPool>
-    ) -> sqlx::Result<Vec<Self>>  {
-        engine.select(LemmaWithKonsepViewOption { lemma: Some(lemma) }).await
+        engine: &DbEngine<sqlx::PgPool>,
+    ) -> sqlx::Result<Vec<Self>> {
+        engine
+            .select(LemmaWithKonsepViewOption { lemma: Some(lemma) })
+            .await
     }
 
     /// Query all lemmas.
-    pub async fn query_all(
-        engine: &DbEngine<sqlx::PgPool>
-    ) -> sqlx::Result<Vec<Self>> {
+    pub async fn query_all(engine: &DbEngine<sqlx::PgPool>) -> sqlx::Result<Vec<Self>> {
         // TODO: Might be a good idea to add limit
         // TODO: And maybe sort by reverse chronology
-        engine.select(LemmaWithKonsepViewOption {lemma: None}).await
+        engine
+            .select(LemmaWithKonsepViewOption { lemma: None })
+            .await
     }
 }
 
@@ -107,123 +113,3 @@ impl IntoViewMap<LemmaWithKonsepView> for Vec<LemmaWithKonsepView> {
             .collect()
     }
 }
-// #[cfg(test)]
-// mod test {
-//     use super::*;
-//
-//     #[sqlx::test(fixtures("lemma", "lemma_2"))]
-//     fn test_lemma_w_konsep_view(pool: sqlx::Pool<sqlx::Sqlite>) {
-//         let views: Vec<LemmaWithKonsepView> = LemmaWithKonsepView::query_all(&pool).await.unwrap();
-//         let mut data = dbg!(LemmaItem::from_views(&views)
-//             .into_iter()
-//             .sorted_by(|a, b| a.lemma.cmp(&b.lemma)));
-//         assert_eq!(
-//             data.next().unwrap(),
-//             LemmaItem {
-//                 id: DbProvided::Known(1),
-//                 lemma: "cakera tokokan".into(),
-//                 konseps: vec![KonsepItem {
-//                     id: DbProvided::Known(1),
-//                     keterangan: "gas-gas dan debu yang mengelilingi lohong hitam".into(),
-//                     golongan_kata: "kata nama".into(),
-//                     cakupans: vec!["Teori Relativiti".into(), "Astrofizik".into(),],
-//                     kata_asing: vec![KataAsingItem {
-//                         nama: "accretion disk".into(),
-//                         bahasa: "english".into(),
-//                     }],
-//                 },],
-//             },
-//         );
-//         assert_eq!(
-//             data.next().unwrap(),
-//             LemmaItem {
-//                 id: DbProvided::Known(2),
-//                 lemma: "ufuk peristiwa".into(),
-//                 konseps: vec![KonsepItem {
-//                     id: DbProvided::Known(2),
-//                     keterangan: "sempadan terluar lohong hitam".into(),
-//                     golongan_kata: "kata nama".into(),
-//                     cakupans: vec!["Teori Relativiti".into(), "Astrofizik".into(),],
-//                     kata_asing: vec![KataAsingItem {
-//                         nama: "event horizon".into(),
-//                         bahasa: "english".into(),
-//                     }],
-//                 },],
-//             }
-//         );
-//     }
-//
-//     #[sqlx::test(fixtures("lemma", "lemma_2"))]
-//     fn test_lemma_w_empty_konsep_query_view(pool: sqlx::Pool<sqlx::Sqlite>) {
-//         let views: Vec<LemmaWithKonsepView> =
-//             LemmaWithKonsepView::query_lemma("cakera tokokan".into(), &pool)
-//                 .await
-//                 .unwrap();
-//         let mut data = dbg!(LemmaItem::from_views(&views)
-//             .into_iter()
-//             .sorted_by(|a, b| a.lemma.cmp(&b.lemma)));
-//         assert_eq!(
-//             data.next(),
-//             Some(LemmaItem {
-//                 id: DbProvided::Known(1),
-//                 lemma: "cakera tokokan".into(),
-//                 konseps: vec![KonsepItem {
-//                     id: DbProvided::Known(1),
-//                     keterangan: "gas-gas dan debu yang mengelilingi lohong hitam".into(),
-//                     golongan_kata: "kata nama".into(),
-//                     cakupans: vec!["Astrofizik".into(), "Teori Relativiti".into(),],
-//                     kata_asing: vec![KataAsingItem {
-//                         nama: "accretion disk".into(),
-//                         bahasa: "english".into(),
-//                     }],
-//                 },],
-//             }),
-//         );
-//     }
-//
-//     #[sqlx::test(fixtures("defaults"))]
-//     fn test_lemma_w_empty_kata_asing(pool: sqlx::Pool<sqlx::Sqlite>) {
-//         let mut item = LemmaItem {
-//             id: DbProvided::Unknown,
-//             lemma: "cakera tokokan".into(),
-//             konseps: vec![KonsepItem {
-//                 id: DbProvided::Unknown,
-//                 keterangan: "gas-gas dan debu yang mengelilingi lohong hitam".into(),
-//                 golongan_kata: "kata nama".into(),
-//                 cakupans: vec!["Astrofizik".into(), "Teori Relativiti".into()],
-//                 kata_asing: vec![],
-//             }],
-//         };
-//         let _ = item.clone().insert_safe(&pool).await.unwrap();
-//         let views: Vec<LemmaWithKonsepView> = LemmaWithKonsepView::query_all(&pool).await.unwrap();
-//         let data = LemmaItem::from_views(&views);
-//         item.id = DbProvided::Known(1);
-//         let k = item.konseps.first_mut().unwrap();
-//         k.id = DbProvided::Known(1);
-//         assert_eq!(data.iter().next(), Some(&item),);
-//     }
-//     #[sqlx::test(fixtures("defaults"))]
-//     fn test_lemma_w_empty_cakupan(pool: sqlx::Pool<sqlx::Sqlite>) {
-//         let mut item = LemmaItem {
-//             id: DbProvided::Unknown,
-//             lemma: "cakera tokokan".into(),
-//             konseps: vec![KonsepItem {
-//                 id: DbProvided::Unknown,
-//                 keterangan: "gas-gas dan debu yang mengelilingi lohong hitam".into(),
-//                 golongan_kata: "kata nama".into(),
-//                 cakupans: vec![],
-//                 kata_asing: vec![KataAsingItem {
-//                     nama: "accretion disk".into(),
-//                     bahasa: "english".into(),
-//                 }],
-//             }],
-//         };
-//         let _ = item.clone().insert_safe(&pool).await.unwrap();
-//         let views: Vec<LemmaWithKonsepView> = LemmaWithKonsepView::query_all(&pool).await.unwrap();
-//         let data = LemmaItem::from_views(&views);
-//         item.id = DbProvided::Known(1);
-//         let k = item.konseps.first_mut().unwrap();
-//         k.id = DbProvided::Known(1);
-//         assert_eq!(data.iter().next(), Some(&item),);
-//     }
-// }
