+++ schema/Cargo.toml
@@ -1,7 +1,12 @@
 [package]
 name = "schema"
-version = "0.1.0"
+version = "1.0.0"
 edition = "2021"
 
 [dependencies]
-serde = { version = "1.0.217", features = ["derive"] }
+itertools = { workspace = true }
+serde = { workspace = true }
+sqlx = { workspace = true }
+
+[dev-dependencies]
+serde_json = "1.0.135"
