+++ schema/src/items.rs
@@ -1,3 +1,6 @@
+//! These items are structures that are often nested
+//! which represent the intended structure of the underlying data.
+
 pub mod cakupan;
 pub mod kata_asing;
 pub mod konsep;
